LPS-98417 removing form field tests from ProjectTemplatesTest and creating them inside of the form field template project
