+++ modules/sdk/project-templates/project-templates/src/test/java/com/liferay/project/templates/ProjectTemplatesTest.java
@@ -660,282 +660,6 @@ public class ProjectTemplatesTest implements BaseProjectTemplatesTestCase {
 	}
 
 	@Test
-	public void testBuildTemplateFormField70() throws Exception {
-		File gradleProjectDir = _buildTemplateWithGradle(
-			"form-field", "foobar", "--liferay-version", "7.0.6");
-
-		testContains(
-			gradleProjectDir, "bnd.bnd", "Bundle-Name: foobar",
-			"Web-ContextPath: /dynamic-data-foobar-form-field");
-		testContains(
-			gradleProjectDir, "build.gradle",
-			"apply plugin: \"com.liferay.plugin\"",
-			DEPENDENCY_PORTAL_KERNEL + ", version: \"2.0.0");
-		testContains(
-			gradleProjectDir,
-			"src/main/java/foobar/form/field/FoobarDDMFormFieldRenderer.java",
-			"property = \"ddm.form.field.type.name=foobar\"",
-			"public class FoobarDDMFormFieldRenderer extends " +
-				"BaseDDMFormFieldRenderer {",
-			"ddm.Foobar", "/META-INF/resources/foobar.soy");
-		testContains(
-			gradleProjectDir,
-			"src/main/java/foobar/form/field/FoobarDDMFormFieldType.java",
-			"ddm.form.field.type.js.class.name=Liferay.DDM.Field.Foobar",
-			"ddm.form.field.type.js.module=foobar-form-field",
-			"ddm.form.field.type.label=foobar-label",
-			"ddm.form.field.type.name=foobar",
-			"public class FoobarDDMFormFieldType extends BaseDDMFormFieldType",
-			"return \"foobar\";");
-		testContains(
-			gradleProjectDir, "src/main/resources/META-INF/resources/config.js",
-			"foobar-group", "'foobar-form-field': {",
-			"path: 'foobar_field.js',", "'foobar-form-field-template': {");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foobar.soy",
-			"{namespace ddm}", "{template .Foobar autoescape",
-			"<div class=\"form-group foobar-form-field\"");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foobar_field.js",
-			"'foobar-form-field',", "var FoobarField",
-			"value: 'foobar-form-field'", "NAME: 'foobar-form-field'",
-			"Liferay.namespace('DDM.Field').Foobar = FoobarField;");
-
-		File mavenProjectDir = buildTemplateWithMaven(
-			temporaryFolder, "form-field", "foobar", "com.test", mavenExecutor,
-			"-DclassName=Foobar", "-Dpackage=foobar", "-DliferayVersion=7.0.6");
-
-		_buildProjects(gradleProjectDir, mavenProjectDir);
-	}
-
-	@Test
-	public void testBuildTemplateFormField71() throws Exception {
-		File gradleProjectDir = _buildTemplateWithGradle(
-			"form-field", "foobar", "--liferay-version", "7.1.3");
-
-		testContains(
-			gradleProjectDir, "bnd.bnd", "Bundle-Name: foobar",
-			"Web-ContextPath: /dynamic-data-foobar-form-field");
-		testContains(
-			gradleProjectDir, "build.gradle",
-			"apply plugin: \"com.liferay.plugin\"",
-			DEPENDENCY_PORTAL_KERNEL + ", version: \"3.0.0");
-		testContains(
-			gradleProjectDir, "package.json",
-			"\"name\": \"dynamic-data-foobar-form-field\"",
-			",foobar_field.js &&");
-		testContains(
-			gradleProjectDir,
-			"src/main/java/foobar/form/field/FoobarDDMFormFieldRenderer.java",
-			"property = \"ddm.form.field.type.name=foobar\"",
-			"public class FoobarDDMFormFieldRenderer extends " +
-				"BaseDDMFormFieldRenderer {",
-			"DDMFoobar.render", "/META-INF/resources/foobar.soy");
-		testContains(
-			gradleProjectDir,
-			"src/main/java/foobar/form/field/FoobarDDMFormFieldType.java",
-			"ddm.form.field.type.description=foobar-description",
-			"ddm.form.field.type.js.class.name=Liferay.DDM.Field.Foobar",
-			"ddm.form.field.type.js.module=foobar-form-field",
-			"ddm.form.field.type.label=foobar-label",
-			"ddm.form.field.type.name=foobar",
-			"public class FoobarDDMFormFieldType extends BaseDDMFormFieldType",
-			"return \"foobar\";");
-		testContains(
-			gradleProjectDir, "src/main/resources/META-INF/resources/config.js",
-			"field-foobar", "'foobar-form-field': {",
-			"path: 'foobar_field.js',");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foobar.soy",
-			"{namespace DDMFoobar}", "variant=\"'foobar'\"",
-			"foobar-form-field");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foobar.es.js",
-			"import templates from './foobar.soy';", "* Foobar Component",
-			"class Foobar extends Component", "Soy.register(Foobar,",
-			"!window.DDMFoobar", "window.DDMFoobar",
-			"window.DDMFoobar.render = Foobar;", "export default Foobar;");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foobar_field.js",
-			"'foobar-form-field',", "var FoobarField",
-			"value: 'foobar-form-field'", "NAME: 'foobar-form-field'",
-			"Liferay.namespace('DDM.Field').Foobar = FoobarField;");
-
-		File mavenProjectDir = buildTemplateWithMaven(
-			temporaryFolder, "form-field", "foobar", "com.test", mavenExecutor,
-			"-DclassName=Foobar", "-Dpackage=foobar", "-DliferayVersion=7.1.3");
-
-		testContains(
-			mavenProjectDir, "bnd.bnd", "-contract: JavaPortlet,JavaServlet");
-
-		_buildProjects(gradleProjectDir, mavenProjectDir);
-	}
-
-	@Test
-	public void testBuildTemplateFormField71WithHyphen() throws Exception {
-		File gradleProjectDir = _buildTemplateWithGradle(
-			"form-field", "foo-bar", "--liferay-version", "7.1.3");
-
-		testContains(
-			gradleProjectDir, "bnd.bnd", "Bundle-Name: foo-bar",
-			"Web-ContextPath: /dynamic-data-foo-bar-form-field");
-		testContains(
-			gradleProjectDir, "build.gradle",
-			"apply plugin: \"com.liferay.plugin\"",
-			DEPENDENCY_PORTAL_KERNEL + ", version: \"3.0.0");
-		testContains(
-			gradleProjectDir, "package.json",
-			"\"name\": \"dynamic-data-foo-bar-form-field\"",
-			",foo-bar_field.js &&");
-		testContains(
-			gradleProjectDir,
-			"src/main/java/foo/bar/form/field/FooBarDDMFormFieldRenderer.java",
-			"property = \"ddm.form.field.type.name=fooBar\"",
-			"public class FooBarDDMFormFieldRenderer extends " +
-				"BaseDDMFormFieldRenderer {",
-			"DDMFooBar.render", "/META-INF/resources/foo-bar.soy");
-		testContains(
-			gradleProjectDir,
-			"src/main/java/foo/bar/form/field/FooBarDDMFormFieldType.java",
-			"ddm.form.field.type.description=foo-bar-description",
-			"ddm.form.field.type.js.class.name=Liferay.DDM.Field.FooBar",
-			"ddm.form.field.type.js.module=foo-bar-form-field",
-			"ddm.form.field.type.label=foo-bar-label",
-			"ddm.form.field.type.name=fooBar",
-			"public class FooBarDDMFormFieldType extends BaseDDMFormFieldType",
-			"return \"fooBar\";");
-		testContains(
-			gradleProjectDir, "src/main/resources/META-INF/resources/config.js",
-			"field-foo-bar", "'foo-bar-form-field': {",
-			"path: 'foo-bar_field.js',");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foo-bar.soy",
-			"{namespace DDMFooBar}", "variant=\"'fooBar'\"",
-			"foo-bar-form-field");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foo-bar.es.js",
-			"import templates from './foo-bar.soy';", "* FooBar Component",
-			"class FooBar extends Component", "Soy.register(FooBar,",
-			"!window.DDMFooBar", "window.DDMFooBar",
-			"window.DDMFooBar.render = FooBar;", "export default FooBar;");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foo-bar_field.js",
-			"'foo-bar-form-field',", "var FooBarField",
-			"value: 'foo-bar-form-field'", "NAME: 'foo-bar-form-field'",
-			"Liferay.namespace('DDM.Field').FooBar = FooBarField;");
-
-		File mavenProjectDir = buildTemplateWithMaven(
-			temporaryFolder, "form-field", "foo-bar", "com.test", mavenExecutor,
-			"-DclassName=FooBar", "-Dpackage=foo.bar",
-			"-DliferayVersion=7.1.3");
-
-		testContains(
-			mavenProjectDir, "bnd.bnd", "-contract: JavaPortlet,JavaServlet");
-
-		_buildProjects(gradleProjectDir, mavenProjectDir);
-	}
-
-	@Test
-	public void testBuildTemplateFormField72() throws Exception {
-		File gradleProjectDir = _buildTemplateWithGradle(
-			"form-field", "foobar", "--liferay-version", "7.2.0");
-
-		testContains(
-			gradleProjectDir, "bnd.bnd", "Bundle-Name: foobar",
-			"Web-ContextPath: /dynamic-data-foobar-form-field");
-		testContains(
-			gradleProjectDir, "build.gradle",
-			"apply plugin: \"com.liferay.plugin\"",
-			DEPENDENCY_PORTAL_KERNEL + ", version: \"4.4.0");
-		testContains(
-			gradleProjectDir,
-			"src/main/java/foobar/form/field/FoobarDDMFormFieldType.java",
-			"ddm.form.field.type.description=foobar-description",
-			"ddm.form.field.type.label=foobar-label",
-			"ddm.form.field.type.name=foobar",
-			"public class FoobarDDMFormFieldType extends BaseDDMFormFieldType",
-			"return \"foobar\";");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foobar.soy",
-			"{namespace Foobar}", "ddm-field-foobar", "form-control foobar");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foobar.es.js",
-			"import './foobarRegister.soy.js';", "* Foobar Component",
-			"import templates from './foobar.soy.js';",
-			"class Foobar extends Component", "Soy.register(Foobar,",
-			"Foobar.STATE", "export default Foobar;");
-
-		File mavenProjectDir = buildTemplateWithMaven(
-			temporaryFolder, "form-field", "foobar", "com.test", mavenExecutor,
-			"-DclassName=Foobar", "-Dpackage=foobar", "-DliferayVersion=7.2.0");
-
-		testContains(
-			mavenProjectDir, "bnd.bnd", "-contract: JavaPortlet,JavaServlet");
-
-		_buildProjects(gradleProjectDir, mavenProjectDir);
-	}
-
-	@Test
-	public void testBuildTemplateFormField72WithHyphen() throws Exception {
-		File gradleProjectDir = _buildTemplateWithGradle(
-			"form-field", "foo-bar", "--liferay-version", "7.2.0");
-
-		testContains(
-			gradleProjectDir, "bnd.bnd", "Bundle-Name: foo-bar",
-			"Web-ContextPath: /dynamic-data-foo-bar-form-field");
-		testContains(
-			gradleProjectDir, "build.gradle",
-			"apply plugin: \"com.liferay.plugin\"",
-			DEPENDENCY_PORTAL_KERNEL + ", version: \"4.4.0");
-		testContains(
-			gradleProjectDir,
-			"src/main/java/foo/bar/form/field/FooBarDDMFormFieldType.java",
-			"ddm.form.field.type.description=foo-bar-description",
-			"ddm.form.field.type.label=foo-bar-label",
-			"ddm.form.field.type.name=fooBar",
-			"public class FooBarDDMFormFieldType extends BaseDDMFormFieldType",
-			"return \"fooBar\";");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foo-bar.soy",
-			"{namespace FooBar}", "ddm-field-foo-bar", "form-control foo-bar");
-		testContains(
-			gradleProjectDir,
-			"src/main/resources/META-INF/resources/foo-bar.es.js",
-			"import './foo-barRegister.soy.js';", "* FooBar Component",
-			"import templates from './foo-bar.soy.js';",
-			"class FooBar extends Component", "Soy.register(FooBar,",
-			"FooBar.STATE", "export default FooBar;");
-
-		File mavenProjectDir = buildTemplateWithMaven(
-			temporaryFolder, "form-field", "foo-bar", "com.test", mavenExecutor,
-			"-DclassName=FooBar", "-Dpackage=foo-bar",
-			"-DliferayVersion=7.2.0");
-
-		testContains(
-			mavenProjectDir, "bnd.bnd", "-contract: JavaPortlet,JavaServlet");
-
-		_buildProjects(gradleProjectDir, mavenProjectDir);
-	}
-
-	@Test
-	public void testBuildTemplateFormFieldInWorkspace() throws Exception {
-		_testBuildTemplateWithWorkspace(
-			"form-field", "foobar", "build/libs/foobar-1.0.0.jar",
-			"--liferay-version", "7.1.3", "--dependency-management-enabled");
-	}
-
-	@Test
 	public void testBuildTemplateFreeMarkerPortlet70() throws Exception {
 		File gradleProjectDir = _testBuildTemplatePortlet70(
 			"freemarker-portlet", "FreeMarkerPortlet", "templates/init.ftl",
